app/functions/youtube/requirements.txt
azure-functions
azure-storage-blob
requests
datetime
youtube_transcript_api
argparse
pandas
openai
dotenv